+++ src/test/java/ufro/dci/filmaffinityfruna/controller/exception_handlers/MovieRestControllerExceptionTest.java
@@ -1,4 +1,4 @@
-package ufro.dci.filmaffinityfruna.controller.exceptionHandlers;
+package ufro.dci.filmaffinityfruna.controller.exception_handlers;
 
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
@@ -8,28 +8,29 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
-import ufro.dci.filmaffinityfruna.controller.DirectorRestController;
-import ufro.dci.filmaffinityfruna.model.entity.DirectorEntity;
-import ufro.dci.filmaffinityfruna.service.DirectorService;
+import ufro.dci.filmaffinityfruna.controller.MovieRestController;
+import ufro.dci.filmaffinityfruna.model.entity.MovieEntity;
+import ufro.dci.filmaffinityfruna.service.MovieService;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 import static org.mockito.Mockito.doThrow;
 
-@WebMvcTest(DirectorRestController.class)
-class DirectorRestControllerExceptionTest {
+@WebMvcTest(MovieRestController.class)
+class MovieRestControllerExceptionTest {
 
     @Autowired
     private MockMvc mockMvc;
 
     @MockBean
-    private DirectorService directorService;
+    private MovieService movieService;
 
     @Test
     void testHandleGeneralException() throws Exception {
-        Mockito.when(directorService.searchByName("Nolan")).thenThrow(new RuntimeException("Error genérico"));
+        Mockito.when(movieService.searchByName("Inception")).thenThrow(new RuntimeException("Error genérico"));
 
-        mockMvc.perform(get("/director/search/Nolan")
+        mockMvc.perform(get("/movie/search")
+                        .param("name", "Inception")
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isInternalServerError())
                 .andExpect(jsonPath("$.message").value("Ocurrió un error inesperado: Error genérico"));
@@ -38,10 +39,10 @@ class DirectorRestControllerExceptionTest {
     @Test
     void testHandleDatabaseException() throws Exception {
         doThrow(new DataIntegrityViolationException("Violación de clave única"))
-                .when(directorService).register(Mockito.any(DirectorEntity.class));
+                .when(movieService).register(Mockito.any(MovieEntity.class));
 
-        mockMvc.perform(post("/director/register")
-                        .content("{\"name\": \"Christopher Nolan\"}")
+        mockMvc.perform(post("/movie/register")
+                        .content("{\"name\": \"Interstellar\"}")
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isConflict())
                 .andExpect(jsonPath("$.message").value("Error de integridad en la base de datos: Violación de clave única"));
@@ -49,7 +50,7 @@ class DirectorRestControllerExceptionTest {
 
     @Test
     void testHandleValidationException() throws Exception {
-        mockMvc.perform(post("/director/register")
+        mockMvc.perform(post("/movie/register")
                         .content("{}")
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isBadRequest())
