+++ src/test/java/ufro/dci/filmaffinityfruna/controller/exception_handlers/UserRestControllerExceptionTest.java
@@ -1,4 +1,4 @@
-package ufro.dci.filmaffinityfruna.controller.exceptionHandlers;
+package ufro.dci.filmaffinityfruna.controller.exception_handlers;
 
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
@@ -8,29 +8,28 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
-import ufro.dci.filmaffinityfruna.controller.ActorRestController;
-import ufro.dci.filmaffinityfruna.model.entity.ActorEntity;
-import ufro.dci.filmaffinityfruna.service.ActorService;
+import ufro.dci.filmaffinityfruna.controller.UserRestController;
+import ufro.dci.filmaffinityfruna.model.entity.UserEntity;
+import ufro.dci.filmaffinityfruna.service.UserService;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 import static org.mockito.Mockito.doThrow;
 
-@WebMvcTest(ActorRestController.class)
-class ActorRestControllerExceptionTest {
+@WebMvcTest(UserRestController.class)
+class UserRestControllerExceptionTest {
 
     @Autowired
     private MockMvc mockMvc;
 
     @MockBean
-    private ActorService actorService;
+    private UserService userService;
 
     @Test
     void testHandleGeneralException() throws Exception {
-        Mockito.when(actorService.searchByName("Leonardo")).thenThrow(new RuntimeException("Error genérico"));
+        Mockito.when(userService.findById(1)).thenThrow(new RuntimeException("Error genérico"));
 
-        mockMvc.perform(get("/actor/search")
-                        .param("name", "Leonardo")
+        mockMvc.perform(get("/user/1")
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isInternalServerError())
                 .andExpect(jsonPath("$.message").value("Ocurrió un error inesperado: Error genérico"));
@@ -39,10 +38,10 @@ class ActorRestControllerExceptionTest {
     @Test
     void testHandleDatabaseException() throws Exception {
         doThrow(new DataIntegrityViolationException("Violación de clave única"))
-                .when(actorService).register(Mockito.any(ActorEntity.class));
+                .when(userService).register(Mockito.any(UserEntity.class));
 
-        mockMvc.perform(post("/actor/register")
-                        .content("{\"name\": \"Leonardo DiCaprio\"}")
+        mockMvc.perform(post("/user/register")
+                        .content("{\"username\": \"johndoe\", \"password\": \"password123\"}")
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isConflict())
                 .andExpect(jsonPath("$.message").value("Error de integridad en la base de datos: Violación de clave única"));
@@ -50,10 +49,10 @@ class ActorRestControllerExceptionTest {
 
     @Test
     void testHandleValidationException() throws Exception {
-        mockMvc.perform(post("/actor/register")
+        mockMvc.perform(post("/user/register")
                         .content("{}")
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isBadRequest())
-                .andExpect(jsonPath("$.name").exists());
+                .andExpect(jsonPath("$.username").exists());
     }
 }
